+++ CHANGELOG.md
@@ -11,11 +11,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
-## [Unreleased]
+## [0.20.1] - 2026-04-06
 
 ### Added
 
 - Added `xsync.Latch`
+- Added `envelopepb.Envelope.IdempotencyKey` field and the
+  `envelopepb.WithIdempotencyKey()` packer option.
 
 ### Changed
 
@@ -471,6 +473,7 @@ which itself includes a large numbers of breaking changes.
 [0.19.18]: https://github.com/dogmatiq/enginekit/releases/v0.19.18
 [0.19.19]: https://github.com/dogmatiq/enginekit/releases/v0.19.19
 [0.20.0]: https://github.com/dogmatiq/enginekit/releases/v0.20.0
+[0.20.1]: https://github.com/dogmatiq/enginekit/releases/v0.20.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD